Rename PipelineStepExtensions to PipelineStepFactoryExtensions Rename PipelineStepsExtensions  to PipelineStepExtensions
